+++ client/package.json
@@ -9,17 +9,18 @@
     "lint": "eslint src --ext js,jsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview",
     "start": "vite"
-    
   },
   "dependencies": {
-    "react": "^18.2.0",
-    "react-bootstrap": "^2.7.0",
+    "@apollo/client": "^3.7.17",
     "bootstrap": "^5.2.3",
     "graphql": "^16.7.1",
-    "@apollo/client": "^3.7.17",
-    "react-dom": "^18.2.0",
     "jwt-decode": "^3.1.2",
-    "react-router-dom": "^6.2.1"
+    "mdb-react-ui-kit": "^6.1.0",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.7.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.2.1",
+    "typeface-roboto": "^1.1.13"
   },
   "devDependencies": {
     "@types/react": "^18.2.14",
