+++ client/public/index.html
@@ -24,9 +24,10 @@
     -->
   <link href="https://fonts.googleapis.com/css2?family=Rubik+Mono+One&family=Rubik:ital,wght@0,400;1,300&display=swap"
     rel="stylesheet">
+  <link href="https://fonts.googleapis.com/css2?family=Indie+Flower&family=Poppins:wght@100&display=swap" 
+    rel="stylesheet">
 
-
-  <title>Google Book Search</title>
+  <title>G.R.O.W</title>
 </head>
 
 <body>
