+++ client/src/App.jsx
@@ -6,6 +6,7 @@ import { setContext } from '@apollo/client/link/context';
 import SearchBooks from './pages/SearchBooks';
 import SavedBooks from './pages/SavedBooks';
 import Navbar from './components/Navbar';
+import Footer from './components/Footer';
 
 
 const httpLink = createHttpLink({
@@ -52,7 +53,11 @@ function App() {
           </Routes>
         </>
       </Router>
+      <div>
+        <Footer />
+      </div>
     </ApolloProvider>
+    
 
   );
 }
