+++ client/src/index.css
@@ -1,6 +1,48 @@
 body {
-  font-family: 'Rubik', sans-serif;
+  font-family: 'Indie Flower', cursive;
+  background-color: #ad6044;
+  display: flex;
+  flex-direction: column;
 }
+
+h1 {
+  font-family: 'Rubik Mono One', sans-serif;
+  font-size: 3rem;
+  font-weight: 400;
+  margin: 0;
+}
+
+h2 {
+  font-family: 'Rubik', sans-serif;
+  font-size: 2rem;
+  font-weight: 400;
+  margin: 0;
+}
+
+#title {
+  font-family: 'Indie Flower', cursive;
+  font-size: 50px;
+  font-weight: 600;
+  color: #177A67;
+}
+#container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  border-radius: 10px;
+  padding: 20px;
+  margin: 20px;
+  width: 50%;
+  box-shadow: 0 0 10px 0 rgba(0,0,0,0.2);
+}
+
+.root {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+}
+
 
 h1,
 h2,
