+++ client/src/main.jsx
@@ -1,8 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import 'bootstrap/dist/css/bootstrap.min.css';
+import 'mdb-react-ui-kit/dist/css/mdb.min.css';
 import './index.css';
 import App from './App';
+
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
